LayoutTests/js/dom/global-function-resolve-expected.txt

CONSOLE MESSAGE: Window 'showModalDialog' function is deprecated and will be removed soon.
Test to make sure cached lookups on the global object are performed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cachedFunctions[0]() is addEventListener
PASS cachedFunctions[0]() is addEventListener
PASS cachedFunctions[1]() is alert
PASS cachedFunctions[1]() is alert
PASS cachedFunctions[2]() is clearInterval
PASS cachedFunctions[2]() is clearInterval
PASS cachedFunctions[3]() is clearTimeout
PASS cachedFunctions[3]() is clearTimeout
PASS cachedFunctions[4]() is clientInformation
PASS cachedFunctions[4]() is clientInformation
PASS cachedFunctions[5]() is closed
PASS cachedFunctions[5]() is closed
PASS cachedFunctions[6]() is confirm
PASS cachedFunctions[6]() is confirm
PASS cachedFunctions[7]() is console
PASS cachedFunctions[7]() is console
PASS cachedFunctions[8]() is crypto
PASS cachedFunctions[8]() is crypto
PASS cachedFunctions[9]() is description
PASS cachedFunctions[9]() is description
PASS cachedFunctions[10]() is devicePixelRatio
PASS cachedFunctions[10]() is devicePixelRatio
PASS cachedFunctions[11]() is dispatchEvent
PASS cachedFunctions[11]() is dispatchEvent
PASS cachedFunctions[12]() is document
PASS cachedFunctions[12]() is document
PASS cachedFunctions[13]() is getComputedStyle
PASS cachedFunctions[13]() is getComputedStyle
PASS cachedFunctions[14]() is getSelection
PASS cachedFunctions[14]() is getSelection
PASS cachedFunctions[15]() is history
PASS cachedFunctions[15]() is history
PASS cachedFunctions[16]() is innerHeight
PASS cachedFunctions[16]() is innerHeight
PASS cachedFunctions[17]() is innerWidth
PASS cachedFunctions[17]() is innerWidth
PASS cachedFunctions[18]() is location
PASS cachedFunctions[18]() is location
PASS cachedFunctions[19]() is locationbar
PASS cachedFunctions[19]() is locationbar
PASS cachedFunctions[20]() is menubar
PASS cachedFunctions[20]() is menubar
PASS cachedFunctions[21]() is moveBy
PASS cachedFunctions[21]() is moveBy
PASS cachedFunctions[22]() is moveTo
PASS cachedFunctions[22]() is moveTo
PASS cachedFunctions[23]() is name
PASS cachedFunctions[23]() is name
PASS cachedFunctions[24]() is navigator
PASS cachedFunctions[24]() is navigator
PASS cachedFunctions[25]() is open
PASS cachedFunctions[25]() is open
PASS cachedFunctions[26]() is opener
PASS cachedFunctions[26]() is opener
PASS cachedFunctions[27]() is outerHeight
PASS cachedFunctions[27]() is outerHeight
PASS cachedFunctions[28]() is outerWidth
PASS cachedFunctions[28]() is outerWidth
PASS cachedFunctions[29]() is pageXOffset
PASS cachedFunctions[29]() is pageXOffset
PASS cachedFunctions[30]() is pageYOffset
PASS cachedFunctions[30]() is pageYOffset
PASS cachedFunctions[31]() is parent
PASS cachedFunctions[31]() is parent
PASS cachedFunctions[32]() is prompt
PASS cachedFunctions[32]() is prompt
PASS cachedFunctions[33]() is releaseEvents
PASS cachedFunctions[33]() is releaseEvents
PASS cachedFunctions[34]() is removeEventListener
PASS cachedFunctions[34]() is removeEventListener
PASS cachedFunctions[35]() is resizeBy
PASS cachedFunctions[35]() is resizeBy
PASS cachedFunctions[36]() is resizeTo
PASS cachedFunctions[36]() is resizeTo
PASS cachedFunctions[37]() is screen
PASS cachedFunctions[37]() is screen
PASS cachedFunctions[38]() is screenLeft
PASS cachedFunctions[38]() is screenLeft
PASS cachedFunctions[39]() is screenTop
PASS cachedFunctions[39]() is screenTop
PASS cachedFunctions[40]() is screenX
PASS cachedFunctions[40]() is screenX
PASS cachedFunctions[41]() is screenY
PASS cachedFunctions[41]() is screenY
PASS cachedFunctions[42]() is scroll
PASS cachedFunctions[42]() is scroll
PASS cachedFunctions[43]() is scrollBy
PASS cachedFunctions[43]() is scrollBy
PASS cachedFunctions[44]() is scrollTo
PASS cachedFunctions[44]() is scrollTo
PASS cachedFunctions[45]() is scrollX
PASS cachedFunctions[45]() is scrollX
PASS cachedFunctions[46]() is scrollY
PASS cachedFunctions[46]() is scrollY
PASS cachedFunctions[47]() is setInterval
PASS cachedFunctions[47]() is setInterval
PASS cachedFunctions[48]() is setTimeout
PASS cachedFunctions[48]() is setTimeout
PASS cachedFunctions[49]() is showModalDialog
PASS cachedFunctions[49]() is showModalDialog
PASS cachedFunctions[50]() is status
PASS cachedFunctions[50]() is status
PASS cachedFunctions[51]() is stop
PASS cachedFunctions[51]() is stop
PASS cachedFunctions[52]() is window
PASS cachedFunctions[52]() is window
PASS successfullyParsed is true

TEST COMPLETE
